backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Henrique Hornos <hhornos@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
FastAPI-SQLAlchemy = "^0.2.1"
python-multipart = "^0.0.5"
pydantic = "^1.9.0"
mysqlclient = "^2.1.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
gunicorn = "^20.1.0"
PyYAML = "^6.0"
stackdriver-formatter = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pylint = "^2.12.2"
pytest-cov = "^3.0.0"
python-dotenv = "^0.19.2"
alembic = "^1.7.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
